pyproject.toml
[project]
name = "eazy-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "sse-starlette>=3.3.4",
    "uvicorn>=0.44.0",
]
